serve/requirements.txt
promptflow
promptflow[azure]
promptflow-tools
python-dotenv

fastapi
uvicorn
pydantic

openai
